pyproject.toml
[project]
name = "ads-txt-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.10.0",
    "duckdb>=1.2.1",
    "fastparquet>=2024.11.0",
    "google-cloud-bigquery[bqstorage,pandas]>=3.30.0",
    "google-cloud-storage>=3.1.0",
    "gspread>=6.2.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scrapy>=2.12.0",
    "tldextract>=5.1.3",
    "truststore>=0.10.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pyarrow>=19.0.1",
]
